rename VQEProgram to VQERuntimeClient
